requirements.txt
openai
chromadb
python-dotenv
tqdm
kivy
kivymd
pandas
